requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
python_dotenv
pandas
azure-storage-file-datalake
azure-identity
pillow
azure-storage-blob
pyarrow